chore: remove ginkgo tests processor
